\name{babel}
\Rdversion{1.1}

\alias{babel}
\alias{babel,ANY,character-method}


\title{
Coerce (convert) spatial objects
}

\description{
this is what it does
}

\usage{babel(x, class, ...)
}

\arguments{
  \item{x}{an object}
  \item{class}{a class name}
  \item{...}{additional arguments}  
}


\value{
An object of the requested class
}


\author{
Barry Rowlinson  and Robert Hijmans
}


\examples{
x = 1:10
}

\keyword{ spatial }


