% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p1table.r
\name{cascade_inner}
\alias{cascade_inner}
\title{internal method that produces sptable() analog from mtable}
\usage{
cascade_inner(x, rcnames = c("v", "bXv", "b"))
}
\description{
internal method that produces sptable() analog from mtable
}

