% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.r
\name{feature_table}
\alias{feature_table}
\title{Normal form for sf}
\usage{
feature_table(x, ...)
}
\arguments{
\item{x}{sf object}

\item{...}{ignored}
}
\description{
A `feature_table` is a normal form for simple features, where all branches are
recorded in one table with attributes object_, branch_, type_, parent_. All instances of parent_
are NA except for the holes in multipolygon.
}
\details{
There is wasted information stored this way, but that's because this is intended as a lowest common denominator format.


There are three tables, objects (the feature attributes and ID), branches (the parts),
coordinates (the X, Y, Z, M values).
}
