% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbabel-package.r
\docType{data}
\name{holey}
\alias{holey}
\title{Multi-part, multi-holed, neighbouring, not completely topological polygons.}
\description{
Created in /data-raw/ from a manual drawing built in Manifold GIS.
}
