% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtable.r
\name{mtable}
\alias{mtable}
\title{Multiple tidy tables}
\usage{
mtable(x, ...)
}
\arguments{
\item{x}{object to tidy}

\item{...}{arguments passed to methods}
}
\value{
list of tibbles
}
\description{
Creates a set of tidy tables from input objects.
}
\examples{
data(holey)
spholey <- sp(holey)
mtable(spholey)
}

