% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.r
\name{primitives}
\alias{primitives}
\alias{primitives.Spatial}
\title{Spatial primitives}
\usage{
primitives(x, ...)

\method{primitives}{Spatial}(x, ...)
}
\arguments{
\item{x}{Spatial object}

\item{...}{other args}
}
\description{
Spatial lines and spatial polygons are composed of 1D primitives (line segments)
Spatial points and spatial multipoints are composed of 0D primtives (coordinates)
Spatial 2D primitives can be generated from 1D primitives either 
- directly (not implemented)
- by triangulation (using ear-clipping or constrained Delaunay)
and from 0D primitives by
- triangulation
}

