% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbabel-package.r
\docType{data}
\name{semap}
\alias{seatt}
\alias{semap}
\title{"South-east" map data.}
\description{
Created in /data-raw/.
}
\examples{
# recreate as sp object
mp <- sp(semap, attr_tab = seatt, crs = "+proj=longlat +ellps=WGS84")

}

