% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtable.r
\name{semi_cascade}
\alias{semi_cascade}
\title{Cascading subset on object for mtable.}
\usage{
semi_cascade(x, ..., tables = c("o", "b", "bXv", "v"))
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
Cascading subset on object for mtable.
}

