% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfFromTable.r
\name{sf}
\alias{sf}
\alias{sf.data.frame}
\title{TBD Convert from dplyr tbl form to simple features.}
\usage{
sf(x, ...)

\method{sf}{data.frame}(x, attr_tab = NULL, crs, ...)
}
\arguments{
\item{x}{tibble as created by \code{\link{sptable}}}

\item{...}{unused}

\item{attr_tab}{remaining data from the attributes}

\item{crs}{projection, defaults to \code{NA_character_}}
}
\value{
sf
}
\description{
Not yet implemented.
}
