% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbabel-package.r
\docType{package}
\name{spbabel-package}
\title{A basis for converting between different types of spatial objects.}
\description{
# @importFrom utils
}
\details{
The spbabel package provides functions to round-trip a Spatial object to a single table and back.
}
\section{I. sptable}{
 an round-trip extension to fortify for Spatial
}

\section{II. Conversion}{

 \tabular{ll}{
 \code{\link{sptable}} \tab create a \code{\link[dplyr]{tbl_df}} from Spatial*DataFrame  \cr
 \code{\link{sptable<-}} \tab modify a Spatial object in-place  \cr
 \code{\link{sp}} \tab create Spatial object from table \cr
 }
}

