% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbabel-package.r
\docType{data}
\name{track}
\alias{track}
\title{Multi-object track with x, y, z, and time. 
Created in /data-raw/track.r}
\description{
Multi-object track with x, y, z, and time. 
Created in /data-raw/track.r
}

