% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_.R
\name{xy_}
\alias{xy_}
\alias{xy_}
\alias{xy_.default}
\alias{xy_}
\alias{xy_.BasicRaster}
\title{Get the xy coordinates.}
\usage{
xy_(x)

\method{xy_}{default}(x)

\method{xy_}{BasicRaster}(x)
}
\description{
A generic version of sp's 'coordinates' to get a centroid.
}
\details{
The default method handles any sp or sf object
}
