% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.r
\name{as_tibble.sfg}
\alias{as_tibble.sfg}
\title{Individual geometries as tibbles.}
\usage{
\method{as_tibble}{sfg}(x)
}
\arguments{
\item{x}{sf geometry of type sfg}
}
\value{
tibble
}
\description{
Individual geometries as tibbles.
}
