% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbabel-package.r
\docType{data}
\name{semap}
\alias{semap}
\alias{seatt}
\title{"South-east" map data.}
\description{
Created in /data-raw/ \code{semap} is the \code{sptable} version of some of maptools 'wrld_simpl' and \code{seatt} is the matching attribute data, linked by 'object_'.

Created in /data-raw/.
}
\examples{
# recreate as sp object
mp <- sp(semap, attr_tab = seatt, crs = "+proj=longlat +ellps=WGS84")

}
