% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spFromTable.r
\name{sp}
\alias{sp}
\alias{sp.data.frame}
\title{Convert from dplyr tbl form to Spatial*DataFrame.}
\usage{
sp(x, ...)

\method{sp}{data.frame}(x, attr_tab = NULL, crs, ...)
}
\arguments{
\item{x}{tibble as created by \code{\link{sptable}}}

\item{...}{unused}

\item{attr_tab}{remaining data from the attributes}

\item{crs}{projection, defaults to \code{NA_character_}}
}
\value{
Spatial*
}
\description{
Convert from dplyr tbl form to Spatial*DataFrame.
}
\examples{
library(dplyr)
semap1 <- semap  \%>\% dplyr::filter(y_ > -89.9999)
sp_obj <- sp(semap1, attr_tab = seatt, crs = "+proj=longlat +ellps=WGS84")
## look, seamless Antarctica!
## library(rgdal); plot(spTransform(sp_obj, "+proj=laea +lat_0=-70"))
}
