% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbabel-package.r
\docType{package}
\name{spbabel-package}
\alias{spbabel-package}
\title{Convert between different types of spatial objects.}
\description{
Facilities for converting between different types of spatial objects, including an in-place method to modify the underlying geometry of 'Spatial' classes using data frame idioms.The spbabel package provides functions to round-trip a Spatial object to a single table and back.
}
\details{
\tabular{ll}{
 \code{\link{sptable<-}} \tab modify a Spatial object in-place  \cr
 \code{\link{sptable}} \tab create a \code{\link[tibble]{tibble}} from Spatial DataFrame object  \cr
 \code{\link{sp}} \tab create Spatial DataFrame object from table \cr
 }
}
